+++ sandbox/urls.py
@@ -1,14 +1,10 @@
-from django.conf.urls import include, url
+from django.conf.urls import url
 from django.contrib import admin
 
-from .views import ping_view
+from . import views
 
 
 urlpatterns = (
-    # Examples:
-    # url(r'^$', 'sandbox.views.home', name='home'),
-    # url(r'^blog/', include('blog.urls')),
-
-    url(r'^ping/$', ping_view, name='ping'),
-    url(r'^admin/', include(admin.site.urls)),
+    url(r'^ping/$', views.ping_view, name='ping'),
+    url(r'^admin/', admin.site.urls),
 )